+++ VCore/Extensions/SwiftUI/View.CornerRadius.swift
@@ -1,10 +1,9 @@
 //
 //  View.CornerRadius.swift
-//  VCore
+//  VComponents
 //
-//  Created by Vakhtang Kontridze on 10/6/21.
+//  Created by Vakhtang Kontridze on 10/28/21.
 //
-
 import SwiftUI
 
 // MARK: - Corner Radius